projects/192-simpsons-RAG/data/requirements.txt
chromadb==0.4.18
openai==1.3.7
python-dotenv==1.0.0
